+++ tilemap_editor/data/view.py
@@ -11,7 +11,9 @@ DEFAULT_TILE_SIZE = 50
 
 _current_map = Map()
 _current_map_path = ""
-camera.set_current(camera.Camera(movement_speed=5, zoom_speed=0.01, min_zoom=-4, zoom_scale=2))
+camera.set_current(camera.Camera(
+    movement_speed=5, zoom_speed=0.01, max_zoom=-4, zoom_scale=2))
+
 
 def load_map(file_path: str):
     global _current_map, _current_map_path
@@ -26,6 +28,7 @@ def save_map_to_path(file_path: str):
     global _current_map_path
     map_loader.save_to_path(_current_map.tiles, file_path)
     _current_map_path = file_path
+
 
 def save_map():
     if _current_map_path != "":
@@ -42,6 +45,7 @@ def get_current_map_name():
 
 
 def blit():
+    """Draw the current view onto the screen."""
     cam = camera.get_current()
     tile_size = DEFAULT_TILE_SIZE * cam.get_zoom_factor()
     vertical_tile_amount = math.ceil(window.WINDOW_HEIGHT / tile_size) + 1
@@ -51,7 +55,7 @@ def blit():
     tiles = _current_map.get_tiles(
         min_x, min_y, horizontal_tile_amount, vertical_tile_amount)
     for i, tile_id in enumerate(tiles):
-        if tile_id != None:
+        if tile_id is not None:
             x_i = min_x + i % horizontal_tile_amount
             y_i = min_y + i // horizontal_tile_amount
             x = x_i * tile_size - cam.x
@@ -59,7 +63,15 @@ def blit():
             tile.blit_at(tile_id, (x, y), tile_size)
 
 
-def tile_coords_from_screen_pos(pos: tuple[float]) -> tuple[int]:
+def tile_coords_from_screen_pos(pos: tuple[int]) -> tuple[int]:
+    """Get tile X & Y from screen pixel coordinates.
+
+    Args:
+        pos: a tuple of integers denoting the screen position in pixels
+
+    Returns:
+        a tuple of ints denoting the tile X/Y position.
+    """
     pos_x, pos_y = pos
     cam = camera.get_current()
     tile_size = DEFAULT_TILE_SIZE * cam.get_zoom_factor()
@@ -71,8 +83,18 @@ def tile_coords_from_screen_pos(pos: tuple[float]) -> tuple[int]:
 
     return (min_x + dist_x // tile_size, min_y + dist_y // tile_size)
 
+
 def get_tile(pos: tuple[int, int]):
+    """Get tile_id at a tile position.
+
+    Args:
+        pos: the tile position as tuple of ints.
+
+    Returns:
+        the tile_id of the tile in question
+    """
     return _current_map.get_tile(pos)
+
 
 def set_tile(pos: tuple[int, int], tile_id):
     _current_map.set_tile(pos, tile_id)